% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.r
\name{fn_glm}
\alias{fn_glm}
\title{Simple package to build a GML classifier for the Bank sales data.}
\usage{
fn_glm()
}
\value{
Saved model file
}
\description{
Simple package to build a GML classifier for the Bank sales data.
}
\examples{
fn_glm()
}
